DM-22700: Include command to add offsets to the lookup tables in ATAOS.
Include commands to apply offset to all axis at the same time cumulatively.
Include commands to reset offsets each axis individually or to all axis at the same time.
Include event with applied offset.
Implement tolerance in applying corrections so we only send commands when really necessary.
Implement `applyAxisOffset` command.
Add correction tolerances for the axis. *Still not implemented* but the idea is that the ATAOS will only issue the correction if the value is above the threshold specified. This is to prevent unnecessary commands from the components.