applied fix to correct issue where atspectrographCorrectionComplete may not get sent depending on how much telescope is moving in elevation.
Fixed FIXME referenced in DM-28681

Modified tests to use persistent offset handling (cmd_poriginOffset) instead of cmd_offsetAzEl.

Modified to force publishing of correction events for all filter/grating/wavelength changes, regardless of if an actual correction is necessary. This is required so scripts will know when the ATAOS has finished all spectrograph corrections since a script does not know if an physical correction is required.

Changed fault method to not be asynchronous.

Added function to lower mirrors when disabling CSC (and corrections are note disabled) and when transitioning to fault state.